+++ CHANGELOG.md
@@ -16,5 +16,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [v0.1.0] - 2024-02-10
 
-[v0.2.0]: https://github.com/Fresh-Advance/Sitemap/compare/v0.1.0...b-7.0.x
+[v0.2.0]: https://github.com/Fresh-Advance/Sitemap/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/Fresh-Advance/Sitemap/compare/03839403...v0.1.0